move ObjModelLoader to environment module
